+++ just_audio_windows/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* [chore]: Only trace every method call in debug builds, and warn once about BufferingProgress ([#64](https://github.com/bdlukaa/just_audio_windows/pull/64))
 * [fix]: `load` returns a Dart error instead of terminating the process on an unsupported source ([#62](https://github.com/bdlukaa/just_audio_windows/pull/62))
 * [fix]: Do not implicitly claim the System Media Transport Controls ([#61](https://github.com/bdlukaa/just_audio_windows/pull/61))